+++ packages/react-shoelaces/src/components/Col.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components'
+import { col } from 'shoelaces'
+
+export const Col = styled.div`${col}`
+
+export default Col