+++ packages/react-shoelaces/src/components/Container.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components'
+import { container } from 'shoelaces'
+
+export const Container = styled.div`${container}`
+
+export default Container