+++ packages/react-shoelaces/src/components/test/Col.test.js
@@ -1,0 +1,7 @@
+import Col from '../Col'
+
+describe('<Col />', () => {
+  it('should render a <Col />', () => {
+    expect(shallow(<Col />)).toMatchSnapshot()
+  })
+})