+++ packages/react-shoelaces/src/components/test/Container.test.js
@@ -1,0 +1,7 @@
+import Container from '../Container'
+
+describe('<Container />', () => {
+  it('should render a <Container />', () => {
+    expect(shallow(<Container />)).toMatchSnapshot()
+  })
+})