+++ packages/react-shoelaces/src/components/test/Flex.test.js
@@ -1,0 +1,7 @@
+import Flex from '../Flex'
+
+describe('<Flex />', () => {
+  it('should render a <Flex />', () => {
+    expect(shallow(<Flex />)).toMatchSnapshot()
+  })
+})