+++ packages/react-shoelaces/src/components/test/Row.test.js
@@ -1,0 +1,7 @@
+import Row from '../Row'
+
+describe('<Row />', () => {
+  it('should render a <Row />', () => {
+    expect(shallow(<Row />)).toMatchSnapshot()
+  })
+})