+++ packages/react-shoelaces/webpack.config.js
@@ -1,0 +1,48 @@
+/* global __dirname, require, module*/
+
+const webpack = require('webpack')
+const UglifyJsPlugin = webpack.optimize.UglifyJsPlugin
+const path = require('path')
+const env = require('yargs').argv.env // use --env with webpack 2
+
+let libraryName = 'react-shoelaces'
+
+let plugins = []
+const outputFile = libraryName + '.js'
+
+if (env === 'build') {
+  plugins.push(new UglifyJsPlugin({ minimize: true }))
+}
+
+const config = {
+  entry: __dirname + '/src/index.js',
+  devtool: 'source-map',
+  output: {
+    path: __dirname + '/lib',
+    filename: outputFile,
+    library: libraryName,
+    libraryTarget: 'umd',
+    umdNamedDefine: true
+  },
+  module: {
+    rules: [
+      {
+        test: /(\.jsx|\.js)$/,
+        loader: 'babel-loader',
+        exclude: /(node_modules|bower_components)/
+      },
+      {
+        test: /(\.jsx|\.js)$/,
+        loader: 'eslint-loader',
+        exclude: /node_modules/
+      }
+    ]
+  },
+  resolve: {
+    modules: [path.resolve('./node_modules'), path.resolve('./src')],
+    extensions: ['.json', '.js']
+  },
+  plugins: plugins
+}
+
+module.exports = config