+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.1.0
+## shoelaces v0.1.0
 ##### _September 23, 2017_
 
 - Exposed existing `min`, `max` and `between` media queries via api
@@ -8,4 +8,17 @@
   - flex properties can be used manually by importing flex : `import flex from 'shoelaces'`
 - Added basic prop validation
 - Fixed an issue where `min-width: 0` would render
+([#5](https://github.com/will-hitchcock/react-shoelaces/pull/5))
+
+
+## react-shoelaces v0.1.0
+##### _September 24, 2017_
+
+- Updated shoelaces to v0.1.0, see the shoelaces ([CHANGELOG](https://github.com/will-hitchcock/shoelaces/blob/master/CHANGELOG.md)) for details
+- Added `<Flex />` component
+
+## react-shoelaces v0.0.3
+##### _September 22, 2017_
+
+- Initial working release
 ([#5](https://github.com/will-hitchcock/react-shoelaces/pull/5))